refactor: move atlan client to atlan.py
- Renamed application_sdk/clients/atlan_client.py to atlan.py for cleaner naming
- Updated import paths in test_atlan_client.py to use new module location
- All tests continue to pass with the refactored structure
